restructurization code
