+++ index.html
@@ -45,6 +45,7 @@
     </div>
 
     <script type="module" src="./src/js/index.js"></script>
+    <script type="module" src="./src/js/ivents.js"></script>
 </body>
 
 </html>
