+++ src/js/declaration.js
@@ -15,5 +15,4 @@ export const buttonNext = document.getElementById("button__next");
 export const volumeBar = document.querySelector(".player-volumeBar");
 export const volume = document.querySelector(".volume");
 
-
-export const faveForm = document.getElementById("wave")
+export const faveForm = document.getElementById("wave");
